+++ lib/lotus/cli.rb
@@ -80,7 +80,9 @@ module Lotus
     end
 
     desc 'generate', 'generates a new action'
-    method_option :path,                desc: 'applications path',        type: :string, default: 'apps'
+    method_option :application_base_url, desc: 'application base url',     type: :string
+    method_option :path,                desc: 'applications path',                                         type: :string,  default: 'apps'
+    method_option :skip_view,           desc: 'skip the creation of view and templates (only for action)', type: :boolean, default: false
     method_option :help, aliases: '-h', desc: 'displays the usage method'
 
     # @since 0.3.0