+++ test/fixtures.rb
@@ -41,6 +41,16 @@ module CoffeeShop
   end
 end
 
+module Reviews
+  class Application < Lotus::Application
+    configure do
+      routes do
+        get '/', to: ->{}, as: :root
+      end
+    end
+  end
+end
+
 class RackApp
   def self.call(env)
     [200, {}, ['Hello from RackApp']]