+++ test/integration/cli/generate_test.rb
@@ -4,8 +4,10 @@ describe 'lotus generate' do
   describe 'action' do
     let(:options)           { '' }
     let(:app_name)          { 'web' }
+    let(:new_app_name)      { 'api' }
     let(:template_engine)   { 'erb' }
     let(:framework_testing) { 'minitest' }
+    let(:klass)             { 'test' }
 
     def create_temporary_dir
       @tmp = Pathname.new(@pwd = Dir.pwd).join('tmp/integration/cli/generate')
@@ -32,6 +34,18 @@ template=#{ template_engine }
       `bundle exec lotus generate action #{ app_name } dashboard#index`
     end
 
+    def generate_action_without_view
+      `bundle exec lotus generate action #{ app_name } dashboard#foo --skip-view`
+    end
+
+    def generate_model
+      `bundle exec lotus generate model #{ klass }`
+    end
+
+    def generate_container
+      `bundle exec lotus generate app #{ new_app_name } #{ new_options }`
+    end
+
     def chdir_to_root
       Dir.chdir(@pwd)
     end
@@ -40,6 +54,7 @@ template=#{ template_engine }
       create_temporary_dir
       generate_application
       generate_action
+      generate_action_without_view
     end
 
     def after
@@ -52,6 +67,67 @@ template=#{ template_engine }
       @root.join('apps/web/templates/dashboard/index.html.erb').must_be  :exist?
       @root.join('spec/web/controllers/dashboard/index_spec.rb').must_be :exist?
       @root.join('spec/web/views/dashboard/index_spec.rb').must_be       :exist?
+    end
+
+    it 'generates an action without view' do
+      @root.join('apps/web/controllers/dashboard/foo.rb').must_be      :exist?
+      @root.join('apps/web/views/dashboard/foo.rb').wont_be            :exist?
+      @root.join('apps/web/templates/dashboard/foo.html.erb').wont_be  :exist?
+      @root.join('spec/web/controllers/dashboard/foo_spec.rb').must_be :exist?
+      @root.join('spec/web/views/dashboard/foo_spec.rb').wont_be       :exist?
+    end
+
+    describe 'when application generates new model' do
+      before do
+        generate_model
+      end
+
+      it 'generates model' do
+        @root.join('lib/delivery/entities/test.rb').must_be                      :exist?
+        @root.join('lib/delivery/repositories/test_repository.rb').must_be       :exist?
+        @root.join('spec/delivery/entities/test_spec.rb').must_be                :exist?
+        @root.join('spec/delivery/repositories/test_repository_spec.rb').must_be :exist?
+      end
+    end
+
+    describe 'when application generates new container' do
+      let(:new_options) { '' }
+
+      before do
+        generate_container
+      end
+
+      it 'generates new container' do
+        @root.join('apps/api/controllers/home/index.rb').must_be             :exist?
+        @root.join('apps/api/views/home/index.rb').must_be                   :exist?
+        @root.join('apps/api/templates/home/index.html.erb').must_be         :exist?
+        @root.join('apps/api/application.rb').must_be                        :exist?
+        @root.join('apps/api/config/routes.rb').must_be                      :exist?
+        @root.join('apps/api/config/mapping.rb').must_be                     :exist?
+        @root.join('apps/api/views/application_layout.rb').must_be           :exist?
+        @root.join('apps/api/templates/application.html.erb').must_be        :exist?
+      end
+
+      it 'inserts configuration files' do
+        content = @root.join('config/environment.rb').read
+        content.must_match %(mount Api::Application, at: '/api')
+        content.must_match %(require_relative '../apps/api/application')
+        content = @root.join('config/.env.development').read
+        content.must_match %(API_DATABASE_URL)
+        content.must_match %(API_SESSIONS_SECRET)
+        content = @root.join('config/.env.test').read
+        content.must_match %(API_DATABASE_URL)
+        content.must_match %(API_SESSIONS_SECRET)
+      end
+
+      describe 'with options application base url' do
+        let(:new_options) { ' --application-base-url=/api_v1' }
+
+        it 'inserts configuration files' do
+          content = @root.join('config/environment.rb').read
+          content.must_match %(mount Api::Application, at: '/api_v1')
+        end
+      end
     end
 
     describe 'when application is generated with minitest' do