+++ test/version_test.rb
@@ -2,6 +2,6 @@ require 'test_helper'
 
 describe Lotus::VERSION do
   it 'returns current version' do
-    Lotus::VERSION.must_equal '0.3.0'
+    Lotus::VERSION.must_equal '0.3.1'
   end
 end